Add deliveries log
