+++ app/assets/stylesheets/webhooks/webhooks.sass
@@ -3,3 +3,9 @@
   a:not(:last-child):after
     content: ","
     padding: 0 1px
+
+.webhooks--delivery-success
+  color: #019875
+
+.webhooks--delivery-error
+  color: #c0392b
