+++ app/cells/webhooks/outgoing/deliveries/row_cell.rb
@@ -1,0 +1,17 @@
+module ::Webhooks
+  module Outgoing
+    module Deliveries
+      class RowCell < ::RowCell
+        include ::IconsHelper
+
+        def log
+          model
+        end
+
+        def time
+          format_time model.updated_at
+        end
+      end
+    end
+  end
+end
