+++ app/views/webhooks/outgoing/admin/index.html.erb
@@ -15,5 +15,5 @@
   </li>
 <% end %>
 
-<%= cell ::Webhooks::Outgoing::TableCell, @webhooks %>
+<%= cell ::Webhooks::Outgoing::Webhooks::TableCell, @webhooks %>
